Conclusão, organização da estrutura de pastas, correçao de problemas e muitas coisas mais

+++ assembler/lib/checker.h
@@ -7,7 +7,7 @@
 #include "string.h"
 
 struct LABEL {
-	int address;
+	long long int address;
     char *name;
     struct LABEL *next;
 };
